pyproject.toml
[project]
name = "phone-relay"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "flask",
    "twilio",
    "python-dotenv",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
]
